move submit button for Troubleshoot tab outside the white box
